feat: Migrate away from next-themes, expose providers individually
