+++ apps/docs/package.json
@@ -16,7 +16,6 @@
     "lint:next": "next lint"
   },
   "dependencies": {
-    "next-themes": "^0.0.15",
     "next": "12.0.7",
     "react-dom": "17.0.2",
     "react": "17.0.2"
