+++ packages/ui/package.json
@@ -37,7 +37,6 @@
     "esno": "^0.14.0",
     "fs-extra": "^10.0.0",
     "lodash": "^4.17.21",
-    "next-themes": "^0.0.15",
     "next": "12.0.7",
     "nodemon": "^2.0.15",
     "npm-run-all": "^4.1.5",
