+++ packages/ui/src/components/Display/Code/CodeEditor.tsx
@@ -2,8 +2,9 @@ import React, { useState, forwardRef, ElementRef, useCallback } from 'react'
 import { LiveEditor } from 'react-live'
 import { noop } from '@aviato/utils'
 
+import { styled } from '~/theme'
 import { useClipboard } from '~/hooks'
-import { styled, useTheme } from '~/theme'
+import { useTheme } from '~/providers'
 import { Tooltip } from '../../Feedback/Tooltip'
 import { CodeLanguage } from './types'
 import { CopyButton } from './CopyButton'
