+++ packages/ui/src/components/Navigation/SideMenu/index.ts
@@ -1,3 +1,2 @@
 export * from './Menu'
-export * from './provider'
 export * from './SideMenu'
