+++ packages/ui/src/providers/index.tsx
@@ -1,1 +1,3 @@
-export * from './combined'
+export * from './aviato'
+export * from './themeProvider'
+export * from './menuProvider'
