+++ packages/ui/src/theme/index.ts
@@ -1,2 +1,1 @@
 export * from './theme'
-export * from './provider'
